Refactor startup sequence
