+++ examples/minimal.rs
@@ -1,10 +1,12 @@
-use bevy::prelude::EventWriter;
+//! This example demonstrates how to use the FmodPlugin to play a sound.
+//! Make sure to follow the instructions in the README.md to set up the demo project.
+
+use bevy::app::{App, Startup};
+use bevy::ecs::event::EventWriter;
 use bevy::DefaultPlugins;
-use bevy::{app::App, prelude::Startup};
+
 use bevy_fmod::{FmodPlugin, PlaySoundEvent};
 
-/// This example demonstrates how to use the FmodPlugin to play a sound.
-/// Make sure to follow the instructions in the README.md to set up the demo project.
 fn main() {
     App::new()
         .add_plugins((
