+++ src/lib.rs
@@ -1,5 +1,6 @@
-mod play_sound_event;
-mod plugin;
-
 pub use crate::play_sound_event::PlaySoundEvent;
 pub use crate::plugin::FmodPlugin;
+
+mod fmod_studio;
+mod play_sound_event;
+mod plugin;
